Rework emoji replacement logic
* Resolves #6
  * Traverse the text nodes containing emoji from back to front to avoid
off by one errors caused by calculating offset lengths. When going back
to front, we hack off the end of the text node that contains an emoji,
replace the emoji in the hacked off chunk, and move on to the remaining
original text node. That is in opposition to and much easier than
traversing the text node from front to back, slicing the text node into
first half, second half, replacing the emoji in the second half, then
trying to figure out what each remaining match index is in relation to
the newly chunked remaining text node.

* additionally, clears up some bad callback logic. popups will no longer
ask for something from a background or content script and wait for it to
come back, it'll just make the request and listen for a response
separately. seems less flakey overall.